+++ packages/core/tests/executors/Serial.test.ts
@@ -1,12 +1,13 @@
 import { mockTool } from '@boost/test-utils';
 import SerialExecutor from '../../src/executors/Serial';
 import Task from '../../src/Task';
+import Context from '../../src/Context';
 
 describe('SerialExecutor', () => {
   let executor: SerialExecutor<any>;
 
   beforeEach(() => {
-    executor = new SerialExecutor(mockTool(), {});
+    executor = new SerialExecutor(mockTool(), new Context());
   });
 
   it('triggers tasks in sequence', async () => {